Relational database implementation - CRUD operations

+++ UniversityEnrollmentApp/UniversityEnrollmentApp/Entities/Candidate.cs
@@ -8,20 +8,20 @@ namespace UniversityEnrollmentApp.Entities
 {
     public class Candidate
     {
-        public int CandidateID { get; set; }
+        public long CandidateID { get; set; }
         public string CandidateFirstName { get; set; }
         public string CandidateLastName { get; set; }
         public DateTime CandidateBirthDate { get; set; }
         public string CandidateAddress { get; set; }
         public bool ApplicationStatus {  get; set; }
-        public int FacultyID { get; set; }
+        public long FacultyID { get; set; }
 
         public Candidate()
         {
             CandidateBirthDate = DateTime.Now;
         }
 
-        public Candidate(int candidateID, string candidateFirstName, string candidateLastName, DateTime candidateBirthDate, string candidateAddress, bool applicationStatus, int facultyID) : this()
+        public Candidate(long candidateID, string candidateFirstName, string candidateLastName, DateTime candidateBirthDate, string candidateAddress, bool applicationStatus, long facultyID) : this()
         {
             CandidateID = candidateID;
             CandidateFirstName = candidateFirstName;
